Add files via upload
The purpose of this Python program is to take a dataset (in CSV format) of extracted OCR field results (e.g. all OCR field data extracted from a set of health claim application documents) and compare each field with another Truth dataset containing the correct values. The program inputs two CSV datasets of the same size, compares them field by field for matches, and outputs a dataset of all mismatching data entries as well as the percentage of accuracy (matches) for each field to an Excel spreadsheet for analysis.